Refactor HTML to be more semantic and simpler
Header, main and footer are now separate.
We also use the flex and classes more appropriately for the scenario.

Key thing was to give the layout less control of how things were displayed.
For e.g. everything being centred in the home page isn't what we want in
the blog list